requirements.txt
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Cors==3.0.10
img2pdf==0.4.0
importlib-metadata==3.1.1
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.5
opencv-python==4.4.0.46
packaging==20.7
pikepdf==2.3.0
Pillow==8.1.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.4.0
